requirements.txt
absl-py==2.1.0
aiocontextvars==0.2.2
aiohttp==3.9.5
aiosignal==1.3.1
anaconda-client==1.2.2
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.15.0
bce-python-sdk==0.9.7
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clyent==1.2.1
cmake==3.29.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contextvars==2.4
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
Cython==3.0.10
dataclasses==0.6
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docker-pycreds==0.4.0
easyocr==1.7.1
email_validator==2.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
immutables==0.20
importlib_metadata==7.1.0
importlib_resources==6.4.0
imutils==0.5.4
install==1.3.5
ipykernel==6.29.4
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
labelImg==1.8.6
lazy_loader==0.4
lmdb==1.4.1
loguru==0.7.2
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nbpresent==3.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.0
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle-gpu==2.6.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
path==16.14.0
path.py==12.5.0
pathtools==0.1.2
pdf2docx==0.5.8
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.1
premailer==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
pyreadline3==3.4.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
rapidfuzz==3.9.0
rarfile==4.2
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-image==0.22.0
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.1.1
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
simplegeneric==0.8.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
terminado==0.18.1
testpath==0.6.0
thop==0.1.1.post2209072238
tifffile==2024.5.3
tinycss2==1.3.0
tomli==2.0.1
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
triton==2.2.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
visualdl==2.5.3
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
win32-setctime==1.1.0
wincertstore==0.2
xlrd==2.0.1
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
